pyproject.toml
[tool.poetry]
name = "process-files"
version = "0.1.0"
description = ""
authors = ["Вахитов Тимур <vt@greentechs.pro>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
celery = "^5.3.5"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
